+++ jsonschema2pojo-core/src/main/java/org/jsonschema2pojo/rules/ConstructorRule.java
@@ -28,8 +28,10 @@ import java.util.Map.Entry;
 import java.util.Objects;
 import java.util.Set;
 import java.util.StringJoiner;
+import java.util.stream.Stream;
 
 import org.apache.commons.lang3.StringUtils;
+import org.apache.commons.lang3.Strings;
 import org.jsonschema2pojo.GenerationConfig;
 import org.jsonschema2pojo.Schema;
 import org.jsonschema2pojo.util.NameHelper;
@@ -47,6 +49,8 @@ import com.sun.codemodel.JInvocation;
 import com.sun.codemodel.JMethod;
 import com.sun.codemodel.JMod;
 import com.sun.codemodel.JVar;
+
+import static java.util.stream.Collectors.toSet;
 
 public class ConstructorRule implements Rule<JDefinedClass, JDefinedClass> {
 
@@ -139,6 +143,10 @@ public class ConstructorRule implements Rule<JDefinedClass, JDefinedClass> {
   }
 
   private void addFieldsConstructor(JDefinedClass instanceClass, Map<String, String> classProperties, Map<String, String> combinedSuperProperties) {
+    if (isConstructorAlreadyAdded(instanceClass, classProperties, combinedSuperProperties)) {
+      return;
+    }
+
     GenerationConfig generationConfig = ruleFactory.getGenerationConfig();
 
     // Generate the constructor with the properties which were located
@@ -151,6 +159,17 @@ public class ConstructorRule implements Rule<JDefinedClass, JDefinedClass> {
 
       generateFieldsBuilderConstructor(baseBuilderClass, concreteBuilderClass, instanceClass, instanceConstructor);
     }
+  }
+
+  private boolean isConstructorAlreadyAdded(JDefinedClass instanceClass, Map<String, String> classProperties, Map<String, String> combinedSuperProperties) {
+    final Set<String> allProperties = Stream.of(classProperties.keySet(), combinedSuperProperties.keySet()).flatMap(Set::stream).collect(toSet());
+    for (Iterator<JMethod> constructorIterator = instanceClass.constructors(); constructorIterator.hasNext(); ) {
+      final Set<String> constructorParams = constructorIterator.next().params().stream().map(JVar::name).collect(toSet());
+      if (constructorParams.equals(allProperties)) {
+        return true;
+      }
+    }
+    return false;
   }
 
   private void addCopyConstructor(JDefinedClass instanceClass, Map<String, String> classProperties, Map<String, String> combinedSuperProperties) {
@@ -194,10 +213,7 @@ public class ConstructorRule implements Rule<JDefinedClass, JDefinedClass> {
     }
 
     NameHelper nameHelper = ruleFactory.getNameHelper();
-    for (Iterator<Entry<String, JsonNode>> properties = node.get("properties")
-        .fields(); properties.hasNext(); ) {
-      Map.Entry<String, JsonNode> property = properties.next();
-
+    for (Map.Entry<String, JsonNode> property : node.get("properties").properties()) {
       JsonNode propertyObj = property.getValue();
       final String javadoc = getJavadocForProperty(propertyObj);
       if (onlyRequired) {
@@ -220,10 +236,10 @@ public class ConstructorRule implements Rule<JDefinedClass, JDefinedClass> {
     private String getJavadocForProperty(JsonNode propertyObj) {
         final StringJoiner stringJoiner = new StringJoiner(" ");
         if (propertyObj.has("title") && StringUtils.isNotBlank(propertyObj.get("title").textValue())) {
-            stringJoiner.add(StringUtils.appendIfMissing(propertyObj.get("title").asText(), "."));
+            stringJoiner.add(Strings.CS.appendIfMissing(propertyObj.get("title").asText(), "."));
         }
         if (propertyObj.has("description") && StringUtils.isNotBlank(propertyObj.get("description").textValue())) {
-            stringJoiner.add(StringUtils.appendIfMissing(propertyObj.get("description").asText(), "."));
+            stringJoiner.add(Strings.CS.appendIfMissing(propertyObj.get("description").asText(), "."));
         }
         return stringJoiner.toString();
     }